Changes
- Moved all server sources into src directory
- Integrated Babel with server
- Removed "config" module and replaced it with ENV
