+++ package.json
@@ -2,15 +2,18 @@
   "name": "seer",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node server",
-    "server": "nodemon server",
+    "prestart": "npm run build",
+    "start": "MODE_ENV=production node ./build/server.js",
+    "server": "nodemon --ignore './client' --exec babel-node src/server.js",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "install-all": "npm install && cd client && npm install && cd ../",
     "pretest": "eslint --ignore-path .eslintignore .",
-    "test": "mocha \"./tests\" --exit",
+    "test": "mocha \"src/tests\" --exit",
+    "prebuild": "rm -rf ./build",
+    "build": "babel -d ./build ./src",
     "heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install && npm install --prefix client && npm run build --prefix client"
   },
   "repository": {
@@ -23,7 +26,7 @@
     "bcryptjs": "^2.4.3",
     "bibtex-parse": "^2.1.0",
     "bootstrap": "^4.4.1",
-    "config": "^3.2.4",
+    "dotenv": "^8.2.0",
     "eslint-plugin-react": "^7.19.0",
     "express": "^4.17.1",
     "express-validator": "^6.3.0",
@@ -34,6 +37,12 @@
     "request": "^2.88.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.6",
+    "@babel/node": "^7.8.7",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/register": "^7.9.0",
     "chai": "^4.2.0",
     "concurrently": "^5.0.1",
     "eslint": "^6.8.0",
