+++ src/config/db.js
@@ -1,7 +1,6 @@
 const mongoose = require("mongoose");
-const config = require("config");
 
-const db = config.get("mongoURI");
+const db = process.env.MONGO_URI;
 
 const connectDB = async () => {
   try {
