+++ src/libs/BibtexParser.js
