+++ src/libs/enums/AffiliationType.js
