+++ src/libs/enums/CategoryType.js
@@ -1,0 +1,6 @@
+
+
+const CategoryType = {
+
+};
+module.exports = CategoryType;
