+++ src/libs/enums/Enum.js
