+++ src/libs/enums/ResearchMethodType.js
