+++ src/libs/enums/SearchConditionType.js
