+++ src/models/QualityRating.js
