+++ src/routes/api/auth.js
@@ -1,28 +1,35 @@
 const express = require("express");
-const gravatar = require("gravatar");
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
-const config = require("config");
 const { check, validationResult } = require("express-validator");
 
+const auth = require("../../middleware/auth");
 const User = require("../../models/User");
 
 const router = express.Router();
 
-// @route    POST api/users
-// @desc     Register user
+// @route    GET api/auth
+// @desc     Test route
+// @access   Public
+router.get("/", auth, async (req, res) => {
+  try {
+    const user = await User.findById(req.user.id).select("-password");
+    res.json(user);
+  }
+  catch (err) {
+    console.error(err.message);
+    res.status(500).send("Server Error");
+  }
+});
+
+// @route    POST api/auth
+// @desc     Authenticate user & get token
 // @access   Public
 router.post(
   "/",
   [
-    check("name", "Name is required")
-      .not()
-      .isEmpty(),
     check("email", "Please include a valid email").isEmail(),
-    check(
-      "password",
-      "Please enter a password with 6 or more characters"
-    ).isLength({ min: 6 })
+    check("password", "Password is required").exists()
   ],
   async (req, res) => {
     const errors = validationResult(req);
@@ -31,34 +38,22 @@ router.post(
       return;
     }
 
-    const { name, email, password } = req.body;
+    const { email, password } = req.body;
 
     try {
-      let user = await User.findOne({ email });
+      const user = await User.findOne({ email });
 
-      if (user) {
-        res.status(400).json({ errors: [{ msg: "User already exists" }] });
+      if (!user) {
+        res.status(400).json({ errors: [{ msg: "Invalid Credentials" }] });
         return;
       }
 
-      const avatar = gravatar.url(email, {
-        s: "200",
-        r: "pg",
-        d: "mm"
-      });
+      const isMatch = await bcrypt.compare(password, user.password);
 
-      user = new User({
-        name,
-        email,
-        avatar,
-        password
-      });
-
-      const salt = await bcrypt.genSalt(10);
-
-      user.password = await bcrypt.hash(password, salt);
-
-      await user.save();
+      if (!isMatch) {
+        res.status(400).json({ errors: [{ msg: "Invalid Credentials" }] });
+        return;
+      }
 
       const payload = {
         user: {
@@ -68,7 +63,7 @@ router.post(
 
       jwt.sign(
         payload,
-        config.get("jwtSecret"),
+        process.env.JWT_SECRET,
         { expiresIn: 360000 },
         (err, token) => {
           if (err) throw err;
