+++ src/server.js
@@ -1,3 +1,5 @@
+require("dotenv").config();
+
 const path = require("path");
 const express = require("express");
 const connectDB = require("./config/db");
