+++ src/tests/BibtexParser.test.js
